Split task and forwarding api tests.
